renamed OverlaySettingsDialogComponent to ImageViewSettingsPanelComponent
