+++ src/components/Dialogs/index.ts
@@ -2,7 +2,6 @@ export * from "./AboutDialog/AboutDialogComponent";
 export * from "./DraggableDialog/DraggableDialogComponent";
 export * from "./ExternalPageDialog/ExternalPageDialogComponent";
 export * from "./FileBrowser/FileBrowserDialogComponent";
-export * from "./OverlaySettings/OverlaySettingsDialogComponent";
 export * from "./TaskProgressDialog/TaskProgressDialogComponent";
 export * from "./RegionDialog/RegionDialogComponent";
 export * from "./PreferenceDialog/PreferenceDialogComponent";
