+++ src/components/ImageView/ImageViewSettingsPanel/ColorComponent.scss
