+++ src/components/ImageView/ImageViewSettingsPanel/ColorComponent.tsx
