+++ src/components/ImageView/ImageViewSettingsPanel/ImageViewSettingsPanelComponent.scss
